requirements.txt
anyio==4.4.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
future==1.0.0
gkeepapi==0.16.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gpsoauth==1.1.1
gspread==6.1.2
h11==0.14.0
halo==0.0.31
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
log-symbols==0.0.14
oauthlib==3.2.2
ollama==0.3.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodomex==3.20.0
pyparsing==3.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
spinners==0.0.24
termcolor==2.4.0
uritemplate==4.1.1
urllib3==1.26.19
